Modificando logica de colisão

+++ src/gameObject.cpp
@@ -7,7 +7,7 @@
 GameObject::GameObject(EntityManager *e, float startingX, float startingY)
 {
   alive = true;
-  pos.set((float)startingX, (float)startingY);
+  pos.set(startingX, startingY);
 }
 
 void GameObject::updatePhysics(float time)
